requirements.txt
numpy
pandas
matplotlib
pyyaml
fastapi
transformers
torch
uvicorn
nltk